refactor(web): reorganize surface cards
