+++ apps/web/src/components/Cards/Bases.tsx
@@ -1,32 +1,5 @@
 import { type FC, type ReactNode } from "react";
 
-type SurfaceCardBaseProps = {
-  children: React.ReactNode;
-  className?: string;
-};
-
-export const CardBase: FC<SurfaceCardBaseProps> = function ({
-  children,
-  className,
-}) {
-  return (
-    <div
-      className={`
-    dark:bg-neutral-850
-    truncate
-    rounded-md
-    border
-    border-border-light
-    p-4
-    dark:border-border-dark
-    ${className}
-    `}
-    >
-      {children}
-    </div>
-  );
-};
-
 type CardTitleBaseProps = {
   children: ReactNode;
   type?: "header" | "footer";
