+++ apps/web/src/components/Cards/SurfaceCards/BlobTransactionCard/Skeleton.tsx
@@ -1,10 +1,10 @@
 import { SkeletonBase, SkeletonRow } from "~/components/SkeletonItems";
-import { CardBase } from "../Bases";
+import { SurfaceCardBase } from "../SurfaceCardBase";
 
 export const BlobTransactionCardSkeleton: React.FC = function () {
   return (
     <SkeletonBase>
-      <CardBase>
+      <SurfaceCardBase>
         <div className="flex justify-between px-3 py-4">
           <div className="flex w-9/12 flex-col space-y-2">
             <SkeletonRow className="h-5 w-10/12" />
@@ -17,7 +17,7 @@ export const BlobTransactionCardSkeleton: React.FC = function () {
             <SkeletonRow className="h-4 w-20" />
           </div>
         </div>
-      </CardBase>
+      </SurfaceCardBase>
     </SkeletonBase>
   );
 };
